move pkg/clusterlabels to pkg/labels
